fix double-init error (non-breaking error)
This was just causing errors to be spat out in the JS console:
basically, initAd() was getting called twice and pushing the same ad
onto the service queue. The second time the service called .defineSlot()
with the same parameters, it does nothing but returns null, which causes
subsequent dereferences to throw.

It was getting called twice because `didEnterViewport` was getting
called before `didInsertElement`.